requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-renderer-xlsx==0.4.0
et-xmlfile==1.0.1
idna==2.10
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
